+++ include/NonPerishableItem.hpp
@@ -6,8 +6,22 @@
 class NonPerishableItem : public Item
 {
 public:
-    NonPerishableItem(const std::string&, const std::string&, unsigned int, unsigned int, double, double, double);
+    /*
+     * Constructs an item with the given values.
+     * @param string Name of item
+     * @param string Category of item (Perishable | NonPerishable)
+     * @param unsigned-long Quantity of item
+     * @param unsigned-long ID of item
+     * @param double Price of item
+     * @param double Cost for store to purchase item
+     * @param double Tax on item
+    */ 
+    NonPerishableItem(const std::string&, const std::string&, unsigned long, unsigned long, double, double, double);
+
+    /* Destructor; does nothing. */
     ~NonPerishableItem() override;
+
+    /* Prints out NonPerishableItem to standard output */
     void print() override;
 private:
     friend class ActiveInventory;
