+++ src/InventoryGenerator.cpp
@@ -8,12 +8,9 @@ InventoryGenerator::InventoryGenerator()
     bad = false;
 }
 
-/*
- * Generates a file named by the file_name param. with entries number of items.
- */
-void InventoryGenerator::generateFile(const std::string& file_name, unsigned int unknown, unsigned int non_perishable, unsigned int perishable)
+void InventoryGenerator::generateFile(const std::string& file_name, unsigned long unknown, unsigned long non_perishable, unsigned long perishable)
 {
-    unsigned int i;
+    unsigned long i;
 
     /* Attempt to open file and throw exception if failed */
     file.open(file_name);
@@ -110,9 +107,6 @@ std::shared_ptr<CSVEntry> InventoryGenerator::generateItem()
     return entry;
 }
 
-/*
- * Generate a non-perishable Item. 
- */
 std::shared_ptr<CSVEntry> InventoryGenerator::generateNonPerishableItem()
 {
     auto entry = generateItem();
@@ -120,9 +114,6 @@ std::shared_ptr<CSVEntry> InventoryGenerator::generateNonPerishableItem()
     return entry;
 }
 
-/*
- * Generate a perishable item.
- */
 std::shared_ptr<CSVEntry> InventoryGenerator::generatePerishableItem()
 {
     std::random_device rand;
@@ -145,18 +136,12 @@ std::shared_ptr<CSVEntry> InventoryGenerator::generatePerishableItem()
     return entry;
 }
 
-/*
- * Tell the generator to set bad vals for this key.
- */
 void InventoryGenerator::setBadKey(const std::string& key)
 {
     bad_keys.insert(key);
     bad = true;
 }
 
-/*
- * Returns whether or not the given key is in the bad key set.
- */
 bool InventoryGenerator::isBadKey(const std::string& key)
 {
     if (random && bad_keys.empty()) return bad;
@@ -164,17 +149,11 @@ bool InventoryGenerator::isBadKey(const std::string& key)
     return !(it == bad_keys.end());
 }
 
-/*
- * Toggles random on/off.
- */
 void InventoryGenerator::toggleRandom()
 {
     random = !random;
 }
 
-/*
- * Sets the random factor.
- */
 void InventoryGenerator::setRandomFactor(const int& i)
 {
     factor = i;
