+++ src/PerishableItem.cpp
@@ -1,16 +1,10 @@
 #include "PerishableItem.hpp"
 
-/*
- * Constructor for PerishableItem; calls base Item constructor, initializes expiration_date member.
-*/
-PerishableItem::PerishableItem(const std::string& nm, const std::string& cat, unsigned int qty, unsigned int idd, double price, double cost, double tx, const std::string& exp)
+PerishableItem::PerishableItem(const std::string& nm, const std::string& cat, unsigned long qty, unsigned long idd, double price, double cost, double tx, const std::string& exp)
 : Item(nm, cat, qty, idd, price, cost, tx), expiration_date(Date(exp))
 {
 }
 
-/*
- * Prints out the item information to standard output.
-*/
 void PerishableItem::print()
 {
     std::cout << "ID: " << id << " Item: " << name << std::endl;
@@ -22,9 +16,6 @@ void PerishableItem::print()
     std::cout << "Profit per unit sold: " << profit << std::endl;
 }
 
-/*
- * Destructor; empty implementation, but needed to shut compiler up.
-*/
 PerishableItem::~PerishableItem()
 {
