+++ src/Sales.cpp
@@ -1,9 +1,6 @@
 #include "Sales.hpp"
 
-/*
- * Constructor for base Sales class, initializes data to given values.
- */
-Sale::Sale(const unsigned int id, const unsigned int sn, const std::string& d, const unsigned int as, const double sp, const double t, const std::string& b, const std::string& s)
+Sale::Sale(const unsigned long id, const unsigned long sn, const std::string& d, const unsigned long as, const double sp, const double t, const std::string& b, const std::string& s)
 : identification(id), sale_number(sn), date(d), amount_sold(as), sale_price(sp), tax(t), buyer(b), seller(s)
 {
     total_price = sale_price + (tax * sale_price);
@@ -16,21 +13,13 @@ Sale::~Sale()
      */
 }
 
-/*
- * Initilizes what file will be used to load/save the sales data to.
- */
 SaleList::SaleList(const std::string& f)
 {
     file_name = f;
     offset = 0;
 }
 
-/*
- * Creates a new sale to put in the sales vector.
- * May change to throwing an exeption in the future.
- * Also may change error checking here based on Sales date restrictions.
- */
-bool SaleList::addSale(const unsigned int id, const unsigned int sn, const std::string& d, const unsigned int as, const double sp, const double t, const std::string& b, const std::string& s)
+bool SaleList::addSale(const unsigned long id, const unsigned long sn, const std::string& d, const unsigned long as, const double sp, const double t, const std::string& b, const std::string& s)
 {
     if(id == 0 || sn == 0 || as == 0 || sp == 0 || d == "" || b == "") 
     {
@@ -43,9 +32,6 @@ bool SaleList::addSale(const unsigned int id, const unsigned int sn, const std::
     return true;
 }
 
-/*
- * Creates a new file with the propper starting format.
- */
 bool SaleList::newFile()
 {
     std::ofstream fout;
@@ -58,19 +44,15 @@ bool SaleList::newFile()
     return true;
 }
 
-/* 
- * Reads in information from given file and holds it in the sales vector
- * If addSale returns false, then an error occured and false is returned.
- */
 bool SaleList::load()
 {
     std::ifstream fin;
     std::string line;
 
-    unsigned int id;
-    unsigned int sn;
+    unsigned long id;
+    unsigned long sn;
     char d[20];
-    unsigned int as;
+    unsigned long as;
     double sp;
     double t;
     char b[50];
@@ -97,14 +79,10 @@ bool SaleList::load()
     return true;
 }
 
-/*
- * Saves the sales information from the vector to the given file, uses append and the offset value 
- * so only newly added information will be saved.
- */
 bool SaleList::save()
 {
     std::ofstream fout;
-    unsigned int i;
+    unsigned long i;
 
     fout.open(file_name.c_str(), std::ios::app);
     if(!fout.is_open()) return false;
